+++ .github/workflows/test.yml
@@ -8,9 +8,9 @@ permissions:
   contents: read
 
 jobs:
-  build:
+  test-jvm:
+    name: Test (JVM)
     runs-on: ubuntu-latest
-
     steps:
       - uses: actions/checkout@v3
       - name: Set up JDK 11
@@ -18,5 +18,21 @@ jobs:
         with:
           java-version: "11"
           distribution: "temurin"
-      - name: Run tests for 3.x
-        run: ./mill "snic[3.3.0]".test
+      - name: Run tests
+        run: ./mill "sline[_]".jvm.test
+
+  test-native:
+    name: Test (Native)
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up JDK 11
+        uses: actions/setup-java@v3
+        with:
+          java-version: "11"
+          distribution: "temurin"
+      - name: Run tests
+        run: ./mill "sline[_]".native.test
