+++ README.md
@@ -1,4 +1,4 @@
-# snic
+# sline
 
 This is a cross-platform library for making interactive CLIs/REPLs in Scala.
 It doesn't do the actual line editing itself: on the JVM, it uses [JLine], and
