+++ jvm/src/sline/JLineCLI.scala
@@ -1,0 +1,3 @@
+package sline
+
+class JLineCLI extends CLI {}
