+++ native/src/sline/ReplxxCLI.scala
@@ -1,0 +1,3 @@
+package sline
+
+class ReplxxCLI extends CLI {}
