+++ shared/src/sline/History.scala
@@ -1,0 +1,3 @@
+package sline
+
+trait History
