+++ shared/src/sline/Sline.scala
@@ -1,0 +1,3 @@
+package sline
+
+case class Sline(val history: History)
